+++ data/users.json
@@ -1,11 +1,1 @@
-[
-  {
-    "nombre": "nombreUsuario1",
-    "apellido": "apellidoUsuario1",
-    "codigoPais": "598",
-    "telefono": "261426142614",
-    "empresa": "Droid SA",
-    "email": "Usuario1@Usuario1.com",
-    "password": "$2b$10$FSyGLEY7B2EngoKWk8cKv.TOIdGG3w.NBTLRFmC91CVdnazRWyy5O"
-  }
-]
+[]
