+++ package.json
@@ -9,6 +9,7 @@
     "cors": "^2.8.5",
     "ejs": "^3.1.6",
     "express": "^4.17.1",
+    "express-session": "^1.17.2",
     "express-validator": "^6.13.0",
     "method-override": "^3.0.0",
     "multer": "^1.4.4"
